requirements.txt
PyYAML
pydantic
typing-extensions

# for cpu temp
# sysctl

liquidctl

humanfriendly
